requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
authlib==1.5.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crcmod==1.7
cryptography==44.0.2
decorator==5.2.1
deprecated==1.2.18
docstring-parser==0.16
editdistance==0.8.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
funasr==1.2.6
gitdb==4.0.12
gitpython==3.1.44
google-adk==0.3.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.90.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.12.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.6.1
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kaldiio==2.18.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markupsafe==3.0.2
mcp==1.6.0
modelscope==1.25.0
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.36.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
omegaconf==2.3.0
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pydeck==0.9.1
pydub==0.25.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-wpe==0.0.1
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
setuptools==80.0.0
shapely==2.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlalchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
streamlit==1.44.1
sympy==1.14.0
tenacity==9.1.2
tensorboardx==2.6.2.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.0
torch-complex==0.4.4
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
litellm==1.67.2
loguru==0.7.3